rename styles file to style and change tabs color #33
