+++ client/src/components/TabNavigator/index.js
@@ -3,7 +3,7 @@ import {View, Text} from 'react-native';
 import EntypoIcon from 'react-native-vector-icons/dist/Entypo';
 import FontAwesome5Icon from 'react-native-vector-icons/dist/FontAwesome5';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
-import {styles} from './styles';
+import {styles} from './style';
 
 // Chat, Contacts and Profile Components for testing the navigation and will be removed later
 function Chat() {
@@ -38,7 +38,7 @@ const TabNavigator = () => {
     <Tab.Navigator
       initialRouteName="Chat"
       tabBarOptions={{
-        activeTintColor: '#51984C',
+        activeTintColor: '#639BBB',
         style: {...styles.tabNavigator},
         labelStyle: {...styles.labelStyle},
       }}>
