+++ client/src/components/TabNavigator/style.js
